+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 1.3.1
+
+#### Date: Jan 8, 2024
+#### Summary of Changes:
+- adds python support for watch_monitoring_result
+
+### 1.3.0
+
+#### Date: Dec 27, 2023
+#### Summary of Changes:
+- adds api for all temporal task routes
+
+
 ### 1.2.6
 
 #### Date: June 13, 2023